AMR/BC: 	modified bc/amr bld script to include handle preproc macro PhaseField_Heat_DBG_DERIVATIVES
Applications:
	All:
		setting non symmetric property to hypre matrix (is symmetric by dflt)
	Heat:
		moved  preproc macro PhaseField_Heat_DBG_DERIVATIVES to Definitions since on it depends HeatProblemTest
		added task_comunicate_before_solve to force any pending communication to be received before using HYPRE (to avoid mixing messages)
		modified logic to ensure correct order of scheduling since checking for coarser/finer can fail if one proc does not have any patch on finest/coarsest level
		fixed refinement computation in task_time_advance_solution_error/task_time_advance_dbg_derivatives_error
		updated additional entries handling in hypresstruct implementations
	PureMetal:
		fixed 3D: postprocessor, epsilon<0 now means rotation around z, z is not modified

DataTypes
	HeatTestProblem is conditioned upon preproc macro PhaseField_Heat_DBG_DERIVATIVES

Util:
	moved  preproc macro PhaseField_Heat_DBG_DERIVATIVES to Definitions.h since on it depends also HeatProblemTest definition
	added util header too Expressions.h

Schedulers:
	DetailedTask: printing also toTasks list
	SchedulerCommon: skipping copy of global variables when specified in applications with overrideVariableBehavior

Solvers
	Stencil7:
		contiguity of members is not assured with optimization: introduced m_value array to fix it and references for compatibility
		removed checks on Stencil7 size in HypreSolver.cc AMR/HypreDriverStruct.cc AMR/HypreDriverSStruct.cc
	HypreSolver:
		implemented missing drivers to HYPRE StructSolvers and added missing parameters to existing ones
		updated readParameters
		changed overrideVariableBehavior to avoid copying hypre_solver structure on regrids

	HypreSStruct:
		ensuring max_levels is defined for FAC
		removed check on being coarser level in scheduleSolve
		removed struct SStructStencil
		added scalar premultiplication operator for MatrixEntry
		added ostream operator for MatrixIndex
		added bld script for generating SStructSolver source for explicit instiantiate class implementations
		moved SStructImplementation/SStructStencils and solver/precond drivers to HYPRE to detail namespace and updated logic, implemented missing drivers
		updated SolverParams and Solver::readParameters
		removed unnecessary ghost dependancies and c2f parameter
		fixed part refinement computation
		updated convergence check
		commented out solver output members converged and residual for TODO implementation
		moved ExtraEntries/ExtraIndices/ExtraValues to detail namespace and fixed bugged range/iterator logic

	solver_specs.xml
		updated list of solver/precond

Implemented ASSERT_NOTHROW and ASSERTEQ_NOTHROW and removed trows from desctuctors in
	CCA/Components/Schedulers/DwDatabase.h
	src/Core/Containers/ConsecutiveRangeSet.cc
	src/Core/Containers/ConsecutiveRangeSet.h
	src/Core/Containers/SuperBox.h
	src/Core/Grid/PatchBVH/PatchBVHNode.cc
	src/Core/Parallel/BufferInfo.cc
	src/Core/Parallel/CrowdMonitor.hpp
	src/Core/Util/InfoMapper.h
	src/Core/Util/RefCounted_gcc.cc

+++ src/Core/Util/InfoMapper.h
@@ -2023,7 +2023,7 @@ public:
 
     // If the key does not exist insert the keys, names, and units.
     if( m_mapInfoMapper.find(key) == m_mapInfoMapper.end() ) {
-      for ( unsigned int i=0; i<m_keys.size(); ++i ) {
+      for ( size_t i=0; i<m_keys.size(); ++i ) {
         m_mapInfoMapper[key].insert( m_keys[i], m_names[i], m_units[i] );
       }
     }
